:recycle: Refactors the Language class and adds a test to detect swift.
